Made database config easier
