+++ debug.py
@@ -1,5 +1,12 @@
 from ssd_api import create_app
 
 if __name__ == '__main__': # pragma: no cover
-    app = create_app()
+    config = {
+        'test': False,
+        'host': 'localhost',
+        'username': 'test_user',
+        'password': 'password',
+        'db_name': 'ssd_sample_database'
+    }
+    app = create_app(**config)
     app.run(debug=True, port=5200)
