+++ ssd_api/db/db_api.py
@@ -1,8 +1,11 @@
 from .Database import Database
 
 
-
 def db_execute(cmd):
+    '''
+    Executes a line of SQL
+    NOTE: Only for reading (SELECT) as it does not commit
+    '''
     with Database.get_db().cursor() as cursor:
         cursor.execute(cmd)
         res = cursor.fetchall()
@@ -10,6 +13,7 @@ def db_execute(cmd):
 
 
 def get_table_names():
+    ''' Get table names in the connected database '''
     with Database.get_db().cursor() as cursor:
         cursor.execute('''SELECT table_name FROM information_schema.tables
         WHERE table_schema = %s;''', Database.db_name)
@@ -18,6 +22,7 @@ def get_table_names():
 
 
 def get_table_columns(table_name):
+    ''' Get the column names of a table '''
     with Database.get_db().cursor() as cursor:
         cursor.execute('''SELECT COLUMN_NAME FROM INFORMATION_SCHEMA.COLUMNS
         WHERE table_name = %s;''', table_name)
